fix: refactor frontend to reorganize app structure
Moved core, layout, shared, and patient-related modules into more appropriate directories for better separation of concerns and maintainability. Updated all relevant imports and paths throughout the application to reflect the new structure. No functional changes were made.
