+++ frontend/src/app/core/guards/auth-guard.spec.ts
