+++ frontend/src/app/layout/components/action-sidebar/action-sidebar.spec.ts
