+++ frontend/src/app/layout/components/footer/footer.spec.ts
