+++ frontend/src/app/layout/components/footer/footer.ts
