+++ frontend/src/app/layout/components/header/header.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from '@angular/core';
-import { LayoutService } from '../../services/layout.service';
-import { HeaderStateService } from '../../services/header-state.service';
+import { LayoutService } from '../../../layout/services/layout.service';
+import { HeaderStateService } from '../../../layout/services/header-state.service';
 import { RouterLink } from '@angular/router';
 
 @Component({
