+++ frontend/src/app/layout/components/sidebar/sidebar.html
