+++ frontend/src/app/layout/components/sidebar/sidebar.spec.ts
