+++ frontend/src/app/layout/components/sidebar/sidebar.ts
@@ -1,7 +1,7 @@
 import { Component, ElementRef, HostListener, inject, signal } from '@angular/core';
 import { Router, RouterLink, RouterLinkActive } from '@angular/router';
-import { LayoutService } from '../../services/layout.service';
-import { AuthService } from '../../services/auth.service';
+import { LayoutService } from '../../../layout/services/layout.service';
+import { AuthService } from '../../../core/services/auth.service';
 
 @Component({
   selector: 'app-sidebar',
