+++ frontend/src/app/layout/services/header-state.service.spec.ts
