+++ frontend/src/app/layout/services/layout.service.ts
