+++ frontend/src/app/modules/patients/components/record-detail/record-detail.spec.ts
