+++ frontend/src/app/modules/patients/services/extraction-api.spec.ts
