+++ frontend/src/app/pages/login/login.ts
@@ -1,7 +1,7 @@
 import { Component, inject } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { Router } from "@angular/router";
-import { AuthService } from "../../services/auth.service";
+import { AuthService } from "../../core/services/auth.service";
 
 @Component({
   selector: "app-login",
