+++ frontend/src/app/pages/profile/profile.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from "@angular/core";
 import { Router } from "@angular/router";
-import { AuthService } from "../../services/auth.service";
+import { AuthService } from "../../core/services/auth.service";
 
 @Component({
   selector: "app-profile",
