+++ frontend/src/app/pages/records/records.ts
@@ -1,8 +1,8 @@
 import { Component, OnDestroy, OnInit, inject } from '@angular/core';
-import { RecordList } from '../../features/record-list/record-list';
-import { RecordListActions } from '../../features/record-list-actions/record-list-actions';
-import { HeaderStateService } from '../../services/header-state.service';
-import { LayoutService } from '../../services/layout.service';
+import { RecordList } from '../../shared/components/data-table/record-list';
+import { RecordListActions } from '../../shared/components/record-list-actions/record-list-actions';
+import { HeaderStateService } from '../../layout/services/header-state.service';
+import { LayoutService } from '../../layout/services/layout.service';
 
 @Component({
   selector: 'app-records',
