+++ frontend/src/app/shared/components/data-table/record-list.html
